% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_date.R
\name{calculate_advice_date}
\alias{calculate_advice_date}
\title{Pick ideal date for next visitation}
\usage{
calculate_advice_date(age, bds_df = NULL)
}
\arguments{
\item{age}{Age of child in decimal years}

\item{bds_df}{\code{data.frame} object with the named columns \code{bds}
containing bds numbers and \code{time} containing decimal years.}
}
\description{
The function picks the best date for a next visitation in the JGZ based on
what has already been done and what still needs doing.
}
\author{
Arjan Huizing, 2023
}
