% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_devlang.R
\name{calculate_advice_devlang}
\alias{calculate_advice_devlang}
\title{Referral advice for language development}
\usage{
calculate_advice_devlang(
  dob = NA_integer_,
  dat_vw41 = NA,
  vw41 = NA,
  dat_vw42 = NA,
  vw42 = NA,
  dat_vw43 = NA,
  vw43 = NA,
  dat_vw44 = NA,
  vw44 = NA,
  dat_vw45 = NA,
  vw45 = NA,
  dat_vw46 = NA,
  vw46 = NA,
  verbose = FALSE,
  force = FALSE
)
}
\arguments{
\item{dob}{Date of birth (ddmmYYYY)}

\item{verbose}{Set to \code{TRUE} to obtain warnings on reference finding.}

\item{y1}{Head circumference at last measurement (cm)}

\item{y0}{Head circumference at previous measurement (cm)}
}
\value{
\code{calculate_advice_devlang} returns an integer, the \code{msgcode}, between
3000-3999.
}
\description{
This function traverses a decision tree for language development based on the
van Wiechen questionnaire.
}
\examples{
msg(calculate_advice_devlang())
msgcode <- calculate_advice_devlang(dob = "01012020",
                                dat_vw41 = "01012022", vw41 = 2,
                                dat_vw42 = "01012022", vw42 = 1)
msg(msgcode)
}
\seealso{
calculate_helpers
}
\author{
Arjan Huizing, Stef van Buuren, 2020
}
