% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_wgt.R
\name{calculate_advice_wgt}
\alias{calculate_advice_wgt}
\title{Referral advice for body weight}
\usage{
calculate_advice_wgt(
  sex = NA_character_,
  dob = as.Date(NA),
  ga = NA,
  dom0 = as.Date(NA),
  y0 = NA,
  dom1 = as.Date(NA),
  y1 = NA,
  hgt1 = NA,
  hgt0 = NA,
  d = NULL,
  useBmi = TRUE
)
}
\arguments{
\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{dob}{Date of birth (class Date)}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{dom0}{Date of previous measurement (Date)}

\item{y0}{Weight at previous measurement (kg)}

\item{dom1}{Date of last measurement (Date)}

\item{y1}{Weight at last measurement (kg)}

\item{hgt1}{Height at last measurement (cm)}

\item{hgt0}{Height at previous measurement (cm)}

\item{d}{Optional, list of derived variables, obtained by
\code{calculate_helpers()}}

\item{useBmi}{Always use bmi cut-off points at ages over 2 (boolean)}
}
\value{
\code{calculate_advice_hgt} returns an integer, the \code{msgcode}
}
\description{
This function traverses the decision tree of the "JGZ-Richtlijn overgewicht
2012".
}
\details{
The decision tree assesses body weight \code{y1} against height \code{hgt1}.
This is done either directly for children below the age of 2, or by
automatically calculating BMI and comparing that against cut-off points
defined in Cole (2000).
}
\examples{
msg(calculate_advice_wgt())
msgcode <- calculate_advice_wgt(sex = "male",
                                dob = as.Date("2015-01-01"),
                                dom1 = as.Date("2018-12-12"),
                                y1 = 25, hgt1 = 120)
msg(msgcode)
}
\seealso{
calculate_helpers
}
\author{
Paula van Dommelen, Stef van Buuren, 2019
}
