% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_birth_z.R
\name{calculate_birth_z}
\alias{calculate_birth_z}
\title{Calculate birth weight SDS relative to Dutch references}
\usage{
calculate_birth_z(y, sex, ga, yname = "wgt", dec = 3)
}
\arguments{
\item{y}{Birth weight (grammes) or birth length (cm).
May be a vector. Converted to numeric.}

\item{sex}{Character, either \code{"M"} (male) or \code{"F"} (female)}

\item{ga}{Gestational age, completed week (Integer or character)}

\item{yname}{Either \code{"wgt"} (for birth weight) or \code{"hgt"}
(for birth length)}

\item{dec}{Number of decimals for rounding}
}
\value{
\preformatted{ Numeric vector of `length(bw)` elements with
         standard deviation scores relative to Dutch birth
         weight references
}
}
\description{
Calculate birth weight SDS relative to Dutch references
}
\examples{
calculate_birth_z(c(2500, 3000), sex = "M", ga = 36)
}
\author{
\preformatted{ Stef van Buuren, 2019
}
}
