% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_helpers.R
\name{calculate_helpers}
\alias{calculate_helpers}
\title{Calculate values required for referral}
\usage{
calculate_helpers(
  yname = "hgt",
  lib = "nl2009",
  sex = NA_character_,
  dob = as.Date(NA),
  bw = NA,
  bl = NA,
  ga = NA,
  etn = NA_character_,
  hgtf = NA,
  hgtm = NA,
  dom1 = as.Date(NA),
  y1 = NA,
  hgt1 = NA,
  dom0 = as.Date(NA),
  y0 = NA,
  hgt0 = NA
)
}
\arguments{
\item{yname}{Character, variable to calculate Z-scores of. Can be one of
\code{"hgt"}, \code{"wgt"}, \code{"hdc"}, \code{"bmi"} or \code{"wfh"}
(weight for height).}

\item{lib}{library to search the reference in}

\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{dob}{Date of birth (class Date)}

\item{bw}{Birth weight (grammes)}

\item{bl}{Birth length (cm)}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{etn}{Etnicity, one of \code{"NL"} (dutch), \code{"TU"} (turkish),
\code{"MA"} (moroccan) or \code{"HS"} (hindustani).}

\item{hgtf}{Height of father (cm)}

\item{hgtm}{Height of mother (cm)}

\item{dom1}{Date of last measurement (Date)}

\item{y1}{\code{y} at last measurement}

\item{hgt1}{\code{hgt} at previous measurement (cm), used when y is
\code{wfh}}

\item{dom0}{Date of previous measurement (Date)}

\item{y0}{\code{y} at previous measurement}

\item{hgt0}{\code{hgt} at last measurement (cm), used when y is \code{wfh}}
}
\value{
\code{calculate_advice_hgt} returns an integer, the \code{msgcode}

\code{calculate_helpers()} returns a \code{list} with the following
elements: \describe{ \item{\code{bw_z}}{Birth weight SDS}
\item{\code{bl_z}}{Birth length SDS} \item{\code{th}}{Target height (cm)}
\item{\code{th_z}}{Target height SDS} \item{\code{age1}}{Age at last
measurement} \item{\code{age0}}{Age at previous measurement}
\item{\code{z1}}{y-value SDS at \code{age1}} \item{\code{z0}}{y-value SDS
at \code{age0}} }
}
\description{
\code{calculate_helpers()} provides an optional pre-calculation for
\code{calculate_advice_hgt()}, \code{calculate_advice_wfh()} and
\code{calculate_advice_hdc()}. The user may wish to divide up calculations
into two steps if intermediate results are needed.
}
\author{
Paula van Dommelen, Stef van Buuren, Arjan Huizing, 2020
}
