% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_msgcode_json.R
\name{convert_msgcode_json}
\alias{convert_msgcode_json}
\title{Pack screener code as JSON string}
\usage{
convert_msgcode_json(url, msgcode, dom1 = ymd(), dom0 = ymd(),
  yname = "hgt")
}
\arguments{
\item{url}{The JAMES site URL created at the request of the client}

\item{msgcode}{Message code, integer}

\item{dom1}{Measurement date 1 (current), class \code{Date}}

\item{dom0}{Measurement date 0 (previous), class \code{Date}}

\item{yname}{Currently only \code{"hgt"}}
}
\value{
A JSON-string
}
\description{
This functions convert the message code to an JSON-string that
can be returned to the client. The format of the JSON conforms
to file \code{file.path(path.package("growthscreener"), "extdata",
"Voorbeeld_antwoordbericht_(2019-08-12).json")}.
}
\author{
Stef van Buuren 2019
}
