% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2age.R
\name{date2age}
\alias{date2age}
\title{Calculate decimal age from date of birth and date of measurement}
\usage{
date2age(dob = NA_character_, dom = NA_character_)
}
\arguments{
\item{dob}{Date of birth}

\item{dom}{Date of measurement}
}
\description{
Calculate decimal age from date of birth and date of measurement
}
\note{
Internal function. Not to be called directly.
}
\examples{
growthscreener:::date2age(dob = "20200217", dom = "20210604")
}
