% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold}
\alias{fold}
\title{Fold long string across multiple lines}
\usage{
fold(s)
}
\arguments{
\item{s}{string}
}
\description{
Fold long string across multiple lines
}
\seealso{
https://stackoverflow.com/questions/2351744/insert-line-breaks-in-long-string-word-wrap
}
