% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetheight.R
\name{hermanussencole}
\alias{hermanussencole}
\title{Target Height Calculation Using the Hermanussen-Cole Method}
\usage{
hermanussencole(
  hgtf,
  hgtm,
  sex,
  rPP = 0.19,
  rPO = 0.58,
  mu = c(183.8, 170.7),
  sigma = c(7.1, 6.3),
  pmu = c(184, 170.6),
  psigma = c(7.1, 6.5)
)
}
\arguments{
\item{hgtf}{Numeric vector with the height of the biological father in cm.}

\item{hgtm}{Numeric vector with the height of the biological mother in cm.
The length of \code{hgtm} must be the same as \code{hgtf}.}

\item{sex}{Character vector indicating the sex of the child for whom the
target height is being calculated. Must be either \code{"male"} or
\code{"female"} and have the same length as \code{hgtf}.}

\item{rPP}{Numeric value for the correlation between the heights of parents.
The default value (0.19) is calculated for the Dutch population.
Hermanussen & Cole (2003) used a value of 0.27.}

\item{rPO}{Numeric value for the correlation between parental height and
offspring height. The default value (0.58) is based on the Dutch
population, while Hermanussen & Cole (2003) used 0.57.}

\item{mu}{Numeric vector of length 2, representing the mean height in cm of
boys and girls in the reference population. The default values \code{mu =
  c(183.8, 170.7)} come from the Fifth Dutch Growth Study (2009).}

\item{sigma}{Numeric vector of length 2, representing the standard deviation
of height in cm for boys and girls in the reference population. The default
values \code{sigma = c(7.1, 6.3)} are based on the Fifth Dutch Growth Study
(2009).}

\item{pmu}{Numeric vector of length 2, representing the mean height in cm of
fathers and mothers in the reference population. The default values
\code{pmu = c(184, 170.6)} are from the Fourth Dutch Growth Study (1997).}

\item{psigma}{Numeric vector of length 2, representing the standard deviation
of height in cm for fathers and mothers in the reference population. The
default values \code{psigma = c(7.1, 6.5)} are from the Fourth Dutch Growth
Study (1997).}
}
\value{
A \code{data.frame} with the three columns:
\describe{
\item{\code{y}}{Target height in cm.}
\item{\code{y.sd}}{Standard deviation of the error term in cm.}
\item{\code{z}}{Target height in standard deviation score (SDS),
relative to the child reference.}
}
}
\description{
This function calculates the target height of a child using the method
proposed by Hermanussen & Cole (2003). The formula accounts for assortative
mating and parent-offspring height correlations, leading to predictions that
are more moderate than the traditional "follow-the-centile" method.
}
\examples{
\dontrun{
# Example 1: Target height for Dutch population, calculated using
# the Hermanussen & Cole method.
hermanussencole(hgtf = 157, hgtm = 154, sex = "male")

# Example 2: The same target height calculated using the simplified method.
targetheight(hgtf = 157, hgtm = 154, sex = "male")

# Example 3: Target height for case discussed in the appendix of
# Hermanussen & Cole (2003).
# A small discrepancy, 166.5 cm (published) versus 166.6 cm (calculated),
# is due to rounding differences.
hermanussencole(hgtf = c(157, 157), hgtm = c(154, 154),
                sex = c("male", "female"),
                rPP = 0.27, rPO = 0.57,
                mu = c(178.2, 163.8), sigma = c(6.8, 6.0),
                pmu = c(178.2, 163.8), psigma = c(6.8, 6.0))
}
}
\references{
Hermanussen M & Cole TJ (2003). "The Calculation of Target Height
Reconsidered." \emph{Hormone Research}, \bold{59}, 180--183.
}
\seealso{
\code{\link{targetheight}}
}
\author{
Stef van Buuren
}
