% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_screeners.R
\name{list_screeners}
\alias{list_screeners}
\title{List the available screeners for JGZ guidelines}
\usage{
list_screeners(ynames = c("hgt", "wgt", "hdc", "lgd"), ...)
}
\arguments{
\item{ynames}{Character vector identifying the measures to be
screened. By default, \code{ynames = c("hgt", "wgt", "hdc")}.}

\item{...}{Not used, but required for extendibility}
}
\value{
A data frame with the following columns
\describe{
\item{\code{Versie}}{Version of \code{growthscreener} package}
\item{\code{yname}}{JAMES name of outcome measure}
\item{\code{Categorie}}{The category of the screening guidelines, 1000/2000/3000/4000}
\item{\code{CategorieOmschrijving}}{A string indicating the measure}
\item{\code{JGZRichtlijn}}{Name of JGZ Richtlijn}
\item{\code{Code}}{Integer message code}
\item{\code{CodeOmschrijving}}{A string with the message}
}
}
\description{
List the available screeners for JGZ guidelines
}
\examples{
screeners <- list_screeners()
table(screeners$Categorie)
}
