% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_curves_ind.R
\name{screen_curves_ind}
\alias{screen_curves_ind}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves_ind(ind)
}
\arguments{
\item{ind}{An object of class \code{minihealth::individual}}
}
\value{
A data frame with 7 columns.
}
\description{
Screen growth curves according to JGZ guidelines
}
\examples{
p <- minihealth::donordata_to_individual(dnr = "smocc", id = 10022)
screen_curves_ind(p)

}
