% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_devlang.R
\name{calculate_advice_devlang}
\alias{calculate_advice_devlang}
\title{Referral advice for language development}
\usage{
calculate_advice_devlang(
  dob = NA_character_,
  dom_vw41 = NA,
  vw41 = NA,
  dom_vw42 = NA,
  vw42 = NA,
  dom_vw43 = NA,
  vw43 = NA,
  dom_vw44 = NA,
  vw44 = NA,
  dom_vw45 = NA,
  vw45 = NA,
  dom_vw46 = NA,
  vw46 = NA,
  verbose = FALSE,
  force = FALSE
)
}
\arguments{
\item{dob}{Date of birth (\code{ddmmYYYY}). Required if \code{dom} is supplied as a
date string.}

\item{dom_vw41}{Date of measurement for van Wiechen item 41. Either a vector
of age in decimal years or a date in the format \code{ddmmYYYY}}

\item{vw41}{Outcome of the van Wiechen item.}
}
\value{
\code{calculate_advice_devlang} returns an integer, the \code{msgcode}, between
4000-4999.
}
\description{
This function traverses the decision tree "JGZ-Richtlijn taalontwikkeling
2012", which is based on the van Wiechen questionnaire for child development.
}
\details{
The observation corresponding to the oldest age is taken is the current
measurement.
}
\examples{
msg(calculate_advice_devlang())
msgcode <- calculate_advice_devlang(dob = "01012020",
                                dom_vw41 = "01012022", vw41 = 2,
                                dom_vw42 = "01012022", vw42 = 1)
msg(msgcode)
}
\author{
Arjan Huizing, Stef van Buuren, 2020
}
