% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_hdc.R
\name{calculate_advice_hdc}
\alias{calculate_advice_hdc}
\title{Referral advice for head circumference}
\usage{
calculate_advice_hdc(
  sex = NA_character_,
  dob = as.Date(NA),
  ga = NA,
  etn = NA_character_,
  dom1 = as.Date(NA),
  y1 = NA,
  dom0 = as.Date(NA),
  y0 = NA,
  d = NULL
)
}
\arguments{
\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{dob}{Date of birth (class Date)}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{etn}{Etnicity, one of \code{"NL"} (dutch), \code{"TU"} (turkish) or
\code{"MA"} (moroccan).}

\item{dom1}{Date of last measurement (Date)}

\item{y1}{Head circumference at last measurement (cm)}

\item{dom0}{Date of previous measurement (Date)}

\item{y0}{Head circumference at previous measurement (cm)}

\item{d}{Optional, list of derived variables, obtained by
\code{calculate_helpers()}}
}
\value{
\code{calculate_advice_hgt} returns an integer, the \code{msgcode}
}
\description{
This function traverses a decision tree for head circumference for children
below the age of 1.
}
\details{
The decision tree always assesses the paired measurements of \code{y1} and
\code{y0}. The last observations (\code{y1}) is taken as the last
measurement, whereas \code{y0} can be one of the previous measurements.
}
\examples{
msg(calculate_advice_hdc())
msgcode <- calculate_advice_hdc(sex = "female", etn = "NL",
                                dob = as.Date("2015-09-01"),
                                dom1 = as.Date("2016-05-01"), y1 = 40,
                                dom0 = as.Date("2015-12-01"), y0 = 36)
msg(msgcode)
}
\seealso{
calculate_helpers
}
\author{
Paula van Dommelen, Stef van Buuren, 2019
}
