% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_hdc.R
\name{calculate_advice_hdc}
\alias{calculate_advice_hdc}
\title{Referral advice for head circumference}
\usage{
calculate_advice_hdc(
  sex = NA_character_,
  ga = NA,
  dom1 = NA_integer_,
  y1 = NA,
  dom0 = NA_integer_,
  y0 = NA,
  test_gain = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{dom1}{Age at last measurement (days)}

\item{y1}{Head circumference at last measurement (cm)}

\item{dom0}{Age at previous measurement (days)}

\item{y0}{Head circumference at previous measurement (cm)}

\item{test_gain}{Logical. Should the increase or decrease in Z-scores be
tested? The default is \code{TRUE}.}

\item{verbose}{Set to \code{TRUE} to obtain warnings on reference finding.}
}
\value{
\code{calculate_advice_hdc} returns an integer, the \code{msgcode},
between 3000-3999.
}
\description{
This function traverses a decision tree for head circumference for children
below the age of 1.
}
\details{
The decision tree assesses both single and paired measurements.
The last observations (\code{y1}) is generally taken as the
last measurement, whereas \code{y0} can be one of the previous
measurements. For more than two measurements, there are many
pairs possible, and these pairs need not be consecutive.
The \code{y0} measurement needs to be defined by the user,
and is informally taken as an earlier measurement that maximizes
the referal probability.
}
\examples{
msg(calculate_advice_hdc())
msgcode <- calculate_advice_hdc(sex = "female",
                                dom1 = 243, y1 = 40,
                                dom0 = 91, y0 = 36)
msg(msgcode)
}
\seealso{
calculate_helpers
}
\author{
Arjan Huizing, Stef van Buuren, 2020
}
