% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_hdc.R
\name{calculate_advice_hdc}
\alias{calculate_advice_hdc}
\title{Referral advice for head circumference}
\usage{
calculate_advice_hdc(
  sex = NA_character_,
  ga = NA,
  dob = NA_character_,
  dom = NA_integer_,
  y = NA,
  test_gain = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{dob}{Date of birth (\code{yyyymmdd}). Required if \code{dom} is supplied as a
date string.}

\item{dom}{Vector with dates of measurements. Either supplied as age in
decimal years or a date in the format \code{yyyymmdd}.}

\item{y}{Head circumference (cm)}

\item{test_gain}{Logical. Should the increase or decrease in Z-scores be
tested? The default is \code{TRUE}.}

\item{verbose}{Set to \code{TRUE} to obtain warnings on reference finding.}
}
\value{
\code{calculate_advice_hdc} returns an integer, the \code{msgcode},
between 3000-3999.
}
\description{
This function traverses a decision tree for head circumference for children
below the age of 1.
}
\details{
The decision tree assesses both single and paired measurements. The
observation corresponding to the oldest age is taken is the current
measurement.
}
\examples{
msg(calculate_advice_hdc())
msgcode <- calculate_advice_hdc(sex = "female",
                                dom = c(0.2491, 0.6653),
                                y = c(36, 40))
msg(msgcode)
}
\author{
Arjan Huizing, Stef van Buuren, 2020
}
