% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_hgt.R
\name{calculate_advice_hgt}
\alias{calculate_advice_hgt}
\title{Referral advice for body height}
\usage{
calculate_advice_hgt(
  sex = NA_character_,
  dob = as.Date(NA),
  bw = NA,
  bl = NA,
  ga = NA,
  etn = "NL",
  hgtf = NA,
  hgtm = NA,
  dom1 = as.Date(NA),
  y1 = NA,
  dom0 = as.Date(NA),
  y0 = NA,
  test_gain = TRUE,
  d = NULL
)
}
\arguments{
\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{dob}{Date of birth (class Date)}

\item{bw}{Birth weight (grammes)}

\item{bl}{Birth length (cm)}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{etn}{Etnicity, one of \code{"NL"} (dutch), \code{"TU"} (turkish),
\code{"MA"} (moroccan) or \code{"HS"} (hindustani).
The default is \code{"NL"}. Only used for target height.}

\item{hgtf}{Height of father (cm)}

\item{hgtm}{Height of mother (cm)}

\item{dom1}{Date of last measurement (Date)}

\item{y1}{Height at last measurement (cm)}

\item{dom0}{Date of previous measurement (Date)}

\item{y0}{Height at previous measurement (cm)}

\item{test_gain}{Logical. Should the increase or decrease in Z-scores be
tested? The default is \code{TRUE}.}

\item{d}{Optional, list of derived variables, obtained by
\code{calculate_helpers()}}
}
\value{
\code{calculate_advice_hgt} returns an integer, the \code{msgcode}
}
\description{
This function traverses the decision tree of the
"JGZ-Richtlijn Lengtegroei 2019".
}
\details{
The decision tree assesses both single and paired measurements.
The last observations (\code{y1}) is generally taken as the
last measurement, whereas \code{y0} can be one of the previous
measurements. For more than two measurements, there are many
pairs possible, and these pairs need not be consecutive.
The \code{y0} measurement needs to be defined by the user,
and is informally taken as an earlier measurement that maximumizes
the referal probability. On the other hand, defining pairs that are
remote in ages (e.g. period between 1 month and 14 years) is probably
not that useful. In practice, we may be interested in setting the
maximum period to, say, five years.
}
\examples{
msg(calculate_advice_hgt())
msgcode <- calculate_advice_hgt(sex = "male",
                                dob = as.Date("2018-07-31"),
                                dom1 = as.Date("2018-12-12"), y1 = 64)
msg(msgcode)
}
\seealso{
calculate_helpers
}
\author{
Paula van Dommelen, Stef van Buuren, 2019
}
