% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_hgt.R
\name{calculate_advice_hgt}
\alias{calculate_advice_hgt}
\title{Referral advice for body height}
\usage{
calculate_advice_hgt(
  sex = NA_character_,
  bw = NA,
  bl = NA,
  ga = NA,
  etn = "NL",
  hgtf = NA,
  hgtm = NA,
  dob = NA_character_,
  dom = NA_character_,
  y = NA,
  y1 = NA,
  y0 = NA,
  dom1 = NA_integer_,
  dom0 = NA_integer_,
  test_gain = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{bw}{Birth weight (grammes)}

\item{bl}{Birth length (cm)}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{etn}{Etnicity, one of \code{"NL"} (dutch), \code{"TU"} (turkish), \code{"MA"}
(moroccan) or \code{"HS"} (hindustani). Currently not functional, always uses
\code{"NL"}. Only used for target height.}

\item{hgtf}{Height of father (cm)}

\item{hgtm}{Height of mother (cm)}

\item{dob}{Date of birth (\code{ddmmYYYY}). Required if \code{dom} is supplied as a
date string.}

\item{dom}{Vector with dates of measurements. Either supplied as age in
decimal years or a date in the format \code{ddmmYYYY}.}

\item{y}{Vector with height measurements (cm)}

\item{y1}{Legacy parameter.}

\item{y0}{Legacy parameter.}

\item{dom1}{Legacy parameter.}

\item{dom0}{Legacy parameter.}

\item{test_gain}{Logical. Should the increase or decrease in Z-scores be
tested? The default is \code{TRUE}.}

\item{verbose}{Set to \code{TRUE} to obtain warnings on reference finding.}
}
\value{
\code{calculate_advice_hgt} returns an integer, the \code{msgcode}
}
\description{
This function traverses the decision tree of the "JGZ-Richtlijn Lengtegroei
2019".
}
\details{
The decision tree assesses both single and paired measurements. The
observation corresponding to the oldest age is taken is the current
measurement.
}
\examples{
msg(calculate_advice_hgt())
msgcode <- calculate_advice_hgt(sex = "male", dob = "01012020",
                                dom = c("01022020", "01062020"),
                                y = c(54, 68),
                                ga = 35,
                                test_gain = FALSE)
msg(msgcode)
}
\author{
Paula van Dommelen, Stef van Buuren, 2021
}
