% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_advice_wgt.R
\name{calculate_advice_wgt}
\alias{calculate_advice_wgt}
\title{Referral advice for weight for height}
\usage{
calculate_advice_wgt(
  sex = NA_character_,
  ga = NA,
  etn = NA,
  dom = NA_integer_,
  y = NA,
  dom_hgt = NA_integer_,
  hgt = NA,
  dob = NA_character_,
  test_gain = TRUE,
  y1 = NA,
  y0 = NA,
  hgt1 = NA,
  hgt0 = NA,
  dom1 = NA_integer_,
  dom0 = NA_integer_,
  verbose = FALSE
)
}
\arguments{
\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{ga}{Gestational age, completed weeks (Integer or character)}

\item{etn}{Etnicity, one of \code{"NL"} (dutch), \code{"TU"} (turkish), \code{"MA"}
(moroccan) or \code{"HS"} (hindustani). Currently not functional, always uses
\code{"NL"}. Only used for target height.}

\item{dom}{Vector with dates of measurements. Either supplied as age in
decimal years or a date in the format \code{yyyymmdd}.}

\item{y}{Vector with height measurements (cm)}

\item{dom_hgt}{Vector with date of measurement relating to height. Either
supplied as age in decimal years or a date in the format \code{yyyymmdd}.}

\item{hgt}{Vector with height measurements (cm)}

\item{dob}{Date of birth (\code{yyyymmdd}). Required if \code{dom} is supplied as a
date string.}

\item{test_gain}{Logical. Should the increase or decrease in Z-scores be
tested? The default is \code{TRUE}.}

\item{y1}{Legacy parameter.}

\item{y0}{Legacy parameter.}

\item{hgt1}{Legacy parameter.}

\item{hgt0}{Legacy parameter.}

\item{dom1}{Legacy parameter.}

\item{dom0}{Legacy parameter.}

\item{verbose}{Set to \code{TRUE} to obtain warnings on reference finding.}
}
\value{
\code{calculate_advice_wgt()} returns an integer, the \code{msgcode}
}
\description{
This function traverses the decision tree of the "JGZ-Richtlijn overgewicht
2012" and "JGZ-Richtlijn Ondergewicht 2019"
}
\details{
The decision tree assesses both single and paired measurements. The
observation corresponding to the oldest age is taken is the current
measurement.
}
\examples{
msg(calculate_advice_wgt())
msgcode <- calculate_advice_wgt(sex = "male", dob = "20200101",
                                dom = c("20200201", "20200601"),
                                dom_hgt = c("20200201", "20200601"),
                                y = c(5.4, 6.8),
                                hgt = c(54, 60),
                                ga = 35,
                                test_gain = FALSE)
msg(msgcode)
}
\author{
Arjan Huizing, Stef van Buuren 2020
}
