% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_th.R
\name{calculate_th}
\alias{calculate_th}
\title{Calculate target height SDS}
\usage{
calculate_th(hgtf, hgtm, sex = NULL, etn = NULL, dec = c(1L, 3L))
}
\arguments{
\item{hgtf}{Length of biological father (cm)}

\item{hgtm}{Length of biological mother (cm)}

\item{sex}{Character, either \code{"male"} or \code{"female"}}

\item{etn}{Etnicity, one of \code{"NL"} (dutch), \code{"TU"} (turkish),
\code{"MA"} (moroccan) or \code{"HS"} (hindustani).}

\item{dec}{Integer vector, length 2, indicating rounding for
th and th_z, respectively}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Numeric, length 2: target height (cm) and target height
         standard deviation score (z), relative to populations
         living in The Netherlands
}\if{html}{\out{</div>}}
}
\description{
Calculate target height SDS
}
\examples{
calculate_th(180, 170, "male", "NL")
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Stef van Buuren, 2019
}\if{html}{\out{</div>}}
}
