% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{ref.nl5defSGAgewicht}
\alias{ref.nl5defSGAlengte}
\alias{nl5.hgt}
\alias{messages}
\alias{bmi_table}
\alias{target}
\title{Table of Dutch birth weight references}
\format{
An object of class \code{data.frame} with 36 rows and 8 columns.

An object of class \code{data.frame} with 36 rows and 8 columns.

An object of class \code{data.frame} with 706 rows and 7 columns.

An object of class \code{data.frame} with 99 rows and 2 columns.

An object of class \code{data.frame} with 132 rows and 6 columns.

An object of class \code{list} of length 2.
}
\usage{
ref.nl5defSGAgewicht

ref.nl5defSGAlengte

nl5.hgt

messages

bmi_table

target
}
\description{
Table of Dutch birth weight references

Table of Dutch birth length references

Table of Dutch nl5 references

Table of return message and return code

Table of BMI cut-off points (Cole, 2000 (overweight), 2007 (thinness), De Wilde)

Example of an object of class \code{target} created by thge \code{bdsreader} package
}
\keyword{datasets}
