% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{ref.nl5defSGAgewicht}
\alias{ref.nl5defSGAlengte}
\alias{nl5.hgt}
\alias{messages_hgt}
\title{Table of Dutch birth weight references}
\format{An object of class \code{data.frame} with 36 rows and 8 columns.}
\usage{
ref.nl5defSGAgewicht

ref.nl5defSGAlengte

nl5.hgt

messages_hgt
}
\description{
Table of Dutch birth weight references

Table of Dutch birth length references

Table of Dutch nl5 references

Table of return message and return code
}
\keyword{datasets}
