% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2age.R
\name{date2age}
\alias{date2age}
\title{Calculate decimal age from date of birth and date of measurement}
\usage{
date2age(dob = NA_character_, dom = NA_character_)
}
\description{
Calculate decimal age from date of birth and date of measurement
}
\note{
Internal function. Not to be called directly.
}
\examples{
growthscreener:::date2age(dob = "17022020", dom = "04062021")
}
