% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Find message string}
\usage{
msg(msgcode)
}
\arguments{
\item{msgcode}{Integer vector with message codes}
}
\value{
A vector of strings with the message code
}
\description{
Find message string
}
\examples{
msg(c(31, 41))
}
