% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_curves_ind.R
\name{screen_curves_ind}
\alias{screen_curves_ind}
\title{Screen growth curves according to JGZ guidelines}
\usage{
screen_curves_ind(
  ind,
  ynames = c("hgt", "wgt", "hdc"),
  na.omit = TRUE,
  recalculate_z = FALSE,
  ...
)
}
\arguments{
\item{ind}{Object of class \code{bdsreader::target}. Support for
\code{minihealth::individual} object is deprecated.}

\item{ynames}{Character vector identifying the measures to be
screened. By default, \code{ynames = c("hgt", "wgt", "hdc")}.}

\item{na.omit}{A logical indicating whether records with a missing
\code{x} (age) or \code{y} (yname) should be removed. Defaults to
\code{TRUE}.}

\item{recalculate_z}{A logical indicating whether Z-scores should be
recalculated. Currently not functional. Z-scores are always
recalculated.}

\item{...}{Not used, but required for extendibility}
}
\value{
A data frame with the following columns
\describe{
\item{\code{Categorie}}{The category of the screening guidelines:
\code{1000 = hgt}, \code{2000 = wgt}, \code{3000 = hdc}}
\item{\code{CategorieOmschrijving}}{A string indicating the measure}
\item{\code{Code}}{Integer message code}
\item{\code{CodeOmschrijving}}{A string with the message}
\item{\code{Versie}}{Version of \code{growthscreener} package}
\item{\code{Leeftijd0}}{First date of the evaluation pair}
\item{\code{Leeftijd1}}{Second date of the evaluation pair}
}
}
\description{
Screen growth curves according to JGZ guidelines
}
\examples{
screen_curves_ind(target)
}
